수집기 및 모델 클래스의 데이터 타입을 개선하고, 서비스 수집기에서 여러 속성을 한 번에 가져오는 기능을 추가하였습니다.
- `ServiceCollector` 클래스의 `GetMultipleProperties` 메서드 시그니처를 수정하여 서비스 객체를 매개변수로 전달하도록 변경하였습니다.
- `CpuCoreInfo`, `DockerContainerInfo`, `NetworkInterface`, `ProcessInfo` 구조체의 데이터 타입을 적절히 수정하여 메모리 사용을 최적화하였습니다.
- `main.cpp`에서 서버 포트 변환 시 타입 캐스팅을 추가하여 코드의 안정성을 높였습니다.

+++ include/models/cpu_info.h
@@ -17,7 +17,7 @@ using namespace std;
  */
 struct CpuCoreInfo
 {
-    int id;            ///< 코어 식별자
+    size_t id;         ///< 코어 식별자
     float usage;       ///< 코어 사용률 (백분율)
     float temperature; ///< 코어 온도 (섭씨)
 };
